requirements.txt
azure-cognitiveservices-vision-computervision==0.7.0
azure-common==1.1.25
certifi==2020.11.8
chardet==3.0.4
colorama==0.4.4
httpie==2.3.0
idna==2.10
isodate==0.6.0
msrest==0.6.19
oauthlib==3.1.0
Pygments==2.7.2
PySocks==1.7.1
requests==2.24.0
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
six==1.15.0
urllib3==1.25.11
